Update MultiThreadDominoes.c
updated Flipping and ready for review. When it prints you need to check match up the indexes for each spot (for the each spot the first pair is the index the 2nd pair is the domino).

I noticed that one time when the second player to play had to draw a domino and could not place it, the hand was reduced to size 1, and the game ended alot quicker.

Also note that the game ends in a tie with all 0's often, because the player have more chances to place their dominos due the perpendicular functionality. However it does still happen that they don't all have 0. Despite it does seem to adhere to the rules of dominos.